refactor: put BasicInvokeStrategy into separate file
